alert pa to el mundo

+++ web/js/controllers/ayudaController.js
@@ -29,7 +29,7 @@ angular.module('GATE')
     bz.respaldar = function () {
 
       servicioDB.backup().then(function (res) {
-
+        swal('Se ha creado el archivo de respaldo de la base de datos! <3');
         console.log(res);
 
       }).catch(function (res) {
@@ -39,12 +39,20 @@ angular.module('GATE')
     }
 
     bz.restaurar = function (F) {
+
       console.log(F);
-      servicioDB.restore(F).then(function (res) {
+      bz.datos.archivoRespaldo = {
+        filename: F
+      }
+      servicioDB.restore(bz.datos.archivoRespaldo).then(function (res) {
         console.log(res);
+        swal('Se ha restaurado la base de datos!');
       }).catch(function (res) {
-        console.log(res);
-      })
+        swal('Ha ocurrido un error!',
+          'No se ha podido restaurar la base de datos',
+          'warning'
+        );
+      });
     }
 
     bz.eliminar = function (index) {
@@ -52,12 +60,31 @@ angular.module('GATE')
         id: bz.datos.respaldos[index].id,
         filename: bz.datos.respaldos[index].filename
       }
+      swal({
+        title: '¿Esta seguro de eliminar el archivo?',
+        text: "esta accion no puede revertirse!",
+        type: 'warning',
+        showCancelButton: true,
+        cancelButtonText: 'Cancelar',
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Si, eliminar!'
+      }).then((result) => {
+        if (result.value) {
 
-      servicioDB.delete(datos).then(function (res) {
-        console.log(res);
-      }).catch(function (res) {
-        console.log(res);
-      })
+          servicioDB.delete(datos).then(function (res) {
+            swal('Se ha eliminado el archivo exitosamente :)');
+          }).catch(function (res) {
+            swal('Ha ocurrido un error!',
+              'No se ha podido eliminar el archivo de respaldo',
+              'error'
+            );
+          })
+        }
+      });
+
+
+
     }
